+++ CHANGELOG.rst
@@ -49,11 +49,13 @@ backwards compatible.
       `fname` instead.
     - This means `.npz` (instead of `numpy`), `.h5` (instead of `h5py`), and
       new `.json`.
-    - New parameter `collect_classes`, which can be used to switch-off the
-      collection of the main classes in root-level dictionaries.
+    - New parameter `collect_classes`, which can be used to switch-on
+      collection of the main classes in root-level dictionaries. By default,
+      they are no longer collected (**changed**).
 
   - `meshes`:
 
+    - `meshes.TensorMesh` **new** inherits from `discretize` if installed.
     - Added `__eq__` to `models.TensorMesh` to compare meshes.
 
 - Dependencies: